Refactor Program class and enhance application instance check
Updated `Program` class to be a partial class and replaced `DllImport` with `LibraryImport` for `SetThreadExecutionState` and `keybd_event` methods. Introduced a mutex in the `Main` method to prevent multiple instances of the application from running simultaneously, with a message box alert for existing instances. Added `<AllowUnsafeBlocks>` property in the project file to enable unsafe code usage.